Break apart parse_beast_header

+++ src/beast/message.rs
@@ -238,7 +238,7 @@ pub enum HeadingSetting {
 
 #[derive(Debug, PartialEq)]
 pub struct Message {
-    pub timestamp: u32,
+    pub timestamp: u64,
     /// RSSI in dBFS
     pub signal_level: f64,
     pub data: Data,
